[#462] refactor: 로그인 문구 파일 추상화
