+++ src/shared/ui/login-required.tsx
